# Add PrecompileTools to speed up time-to-first-solve