--- a/Project.toml
+++ b/Project.toml
@@ -11,6 +11,7 @@ JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Memento = "f28f55f0-a522-5efc-85c2-fe41dfb9b2d9"
 NLsolve = "2774e3e8-f4cf-5e23-947b-6d7e65073b56"
+PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
 [compat]
@@ -22,6 +23,7 @@ JuMP = "1.15"
 Juniper = "~0.9"
 Memento = "~1.0, ~1.1, ~1.2, ~1.3, ~1.4"
 NLsolve = "4.0"
+PrecompileTools = "1"
 SCS = "~0.9, ~1.0, ~2.0"
 julia = "1.6"