--- a/src/PowerModels.jl
+++ b/src/PowerModels.jl
@@ -7,6 +7,7 @@ import JuMP
 import LinearAlgebra
 import Memento
 import NLsolve
+import PrecompileTools
 import SparseArrays
 
 # Create our module level logger (this will get precompiled)
@@ -83,6 +84,22 @@ include("util/flow_limit_cuts.jl")
 
 # this must come last to support automated export
 include("core/export.jl")
+
+PrecompileTools.@setup_workload begin
+    case3 = joinpath(dirname(@__DIR__), "test/data/matpower/case3.m")
+    case9 = joinpath(dirname(@__DIR__), "test/data/matpower/case9.m")
+    PrecompileTools.@compile_workload begin
+        for case in [case3, case9]
+            data = parse_file(case3)
+            _ = instantiate_model(data, ACPPowerModel, build_opf)
+            _ = instantiate_model(data, ACPPowerModel, build_pf)
+            _ = instantiate_model(data, DCPPowerModel, build_opf)
+            _ = instantiate_model(data, DCPPowerModel, build_pf)
+        end
+        _ = compute_ac_pf(case9)
+        _ = compute_dc_pf(case9)
+    end
+end
 
 # Deprecations to be removed in the next breaking release