@ccoffrin 2025-04-25
Nice! @noahrhodes and I need to think a bit about the exact set of stuff that should go in `@setup_workload`. We can merge this now and open an issue to expand the list; or we can hold this open for a ~1 week in the hopes I will find some time to work it.